修复 seaki-pipeline 中 u64 截断与 schema hash 重复计算问题
M2 审核发现 cost.rs 中 u64→u32 转换导致大值静默截断，compiler 中对同一
command 的 manifest 做两次 registry.inspect() 造成重复计算。

cost.rs 将 f64::from(est as u32) 改为 est as f64 / act as f64，f64 精确
范围覆盖到 2^53，足以表示实际 token/内存/CPU 值。

compiler.rs 提取 validate_manifest_hash() 共享辅助函数消除代码重复；
compile() 在预验证阶段将 resource_quota 和 schema_hash 缓存到 HashMap，
后续构建 CompiledStep 时直接复用，避免第二次 registry.inspect()。

Constraint: u64→u32 截断在 token 计数超过 40 亿时必现，LLM 长上下文场景下
容易触发
Constraint: schema hash 重复计算在复杂 pipeline 中放大延迟，每步多一次
registry 查询
Rejected: 使用 wrapping_add / saturating 处理截断 | 错误率仍然不可接受
Rejected: 在 registry 层加内部缓存 | 修改范围过大，当前仅 compiler 需要
Confidence: high
Scope-risk: narrow
Directive: cost.rs 中所有数值比较统一使用 as f64，禁止再出现 as u32 转换
Tested: cargo test -p seaki-pipeline 43 passed，含大值截断边界测试
Tested: cargo fmt + clippy 无警告
Not-tested: 真实 LLM 返回 100 亿+ token 的端到端场景（模拟数据已覆盖）

+++ crates/seaki-pipeline/src/cost.rs
@@ -47,7 +47,7 @@ impl CostEstimate {
                     Err(format!("{label}: actual is 0 but estimate is {est}"))
                 };
             }
-            let ratio = f64::from(est as u32) / f64::from(act as u32);
+            let ratio = est as f64 / act as f64;
             if !(0.5..=2.0).contains(&ratio) {
                 return Err(format!(
                     "{label}: estimate {est} vs actual {act} (ratio {ratio:.2} outside 0.5x~2x)"
